+++ utils/hashPassword.go
@@ -1,4 +1,4 @@
-package util
+package utils
 
 import "golang.org/x/crypto/bcrypt"
