+++ utils/jwt.go
@@ -1,4 +1,4 @@
-package util
+package utils
 
 import (
 	"echo-app/config"
