+++ validation/category.go
@@ -2,7 +2,7 @@ package validation
 
 import (
 	"echo-app/model"
-	"echo-app/util"
+	"echo-app/utils"
 
 	"github.com/labstack/echo/v4"
 )
@@ -16,13 +16,13 @@ func (c Category) CreateBody(next echo.HandlerFunc) echo.HandlerFunc {
 		// bind request data
 		if err := c.Bind(&body); err != nil {
 			if err != nil {
-				return util.Response400(c, nil, err.Error())
+				return utils.Response400(c, nil, err.Error())
 			}
 		}
 
 		// validate
 		if err := body.Validate(); err != nil {
-			return util.Response400(c, nil, err.Error())
+			return utils.Response400(c, nil, err.Error())
 		}
 
 		c.Set("body", body)
@@ -38,13 +38,13 @@ func (c Category) UpdateBody(next echo.HandlerFunc) echo.HandlerFunc {
 		// bind request data
 		if err := c.Bind(&body); err != nil {
 			if err != nil {
-				return util.Response400(c, nil, err.Error())
+				return utils.Response400(c, nil, err.Error())
 			}
 		}
 
 		// validate
 		if err := body.Validate(); err != nil {
-			return util.Response400(c, nil, err.Error())
+			return utils.Response400(c, nil, err.Error())
 		}
 
 		c.Set("body", body)
